Work on optimizing restore
